pyproject.toml
[tool.poetry]
name = "projeto-gen-ai"
version = "0.1.0"
description = ""
authors = ["Thiago <iamthiagovillani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
pydentic = "^0.0.1.dev3"
pydantic = {extras = ["email"], version = "^2.9.2"}
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.38"
mkdocstrings = "^0.26.1"
mkdocstrings-python = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
